+++ CHANGELOG.md
@@ -13,7 +13,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Option to manually load RNA MULTIQC report using the command `scout load report qc_rna`
 - Load RNA MULTIQC automatically for a case if config file contains the `multiqc_rna` key/value
 ### Changed
-- Deprecated CLI commands `scout load <delivery_report, gene_fusion_report, coverage_qc_report, cnv_report>` to replace them with command `scout load report <report type>`
+- Deprecated CLI commands `scout load <delivery_report, gene_fusion_report, coverage_qc_report, cnv_report>` to replace them with command `scout load report -t <report type>`
 ### Fixed
 - Non-admin users saving institute settings would clear loqusdb instance selection